Add plans and cohorts put get
