+++ src/app/views/admissions/student-form/StudentCohorts.jsx
@@ -36,36 +36,11 @@ const propTypes = {
   stdId: PropTypes.number.isRequired,
 };
 
-const actionController = {
-  message: {
-    educational_status: "Educational Status",
-    finantial_status: "Finantial Status",
-    role: "Cohort Role",
-    subscriptions_status: "Subscription Status",
-    plan_financing_status: "Plan Financing Status",
-  },
-  options: {
-    educational_status: [
-      "ACTIVE",
-      "POSTPONED",
-      "SUSPENDED",
-      "GRADUATED",
-      "DROPPED",
-      "NOT_COMPLETING",
-    ],
-    finantial_status: ["FULLY_PAID", "UP_TO_DATE", "LATE"],
-    role: ["TEACHER", "ASSISTANT", "REVIEWER", "STUDENT"],
-    subscriptions_status: ["ACTIVE", "ERROR", "FREE_TRIAL"],
-    plan_financing_status: ["FREE_TRIAL", "ACTIVE", "ERROR"],
-  },
-};
-
 const StudentCohorts = ({ stdId, setCohortOptions }) => {
   const [setMsg] = useState({ alert: false, type: "", text: "" });
   const [isLoading, setIsLoading] = useState(false);
   const [stdCohorts, setStdCohorts] = useState([]);
   const [currentStd, setCurrentStd] = useState({});
-  const [openRoleDialog, setRoleDialog] = useState(false);
   const [openPlanDialog, setOpenPlanDialog] = useState(false);
   const [openDialog, setOpenDialog] = useState(false);
   const [cohort, setCohort] = useState(null);
@@ -94,26 +69,6 @@ const StudentCohorts = ({ stdId, setCohortOptions }) => {
           setStdCohorts(data);
           setCohortOptions(data);
         }
-      })
-      .catch((error) => error);
-  };
-
-  const changeStudentStatus = (value, name, studentId, i) => {
-    const sStatus = {
-      role: stdCohorts[i].role.toUpperCase(),
-      finantial_status: stdCohorts[i].finantial_status,
-      educational_status: stdCohorts[i].educational_status,
-      subscriptions_status: stdCohorts[i].subscriptions_status,
-      plan_financing_status: stdCohorts[i].plan_financing_status,
-      [name]: value,
-    };
-    console.log("SSSSSSTATTTTUSSSS", sStatus);
-
-    bc.admissions()
-      .updateCohortUserInfo(stdCohorts[i].cohort.id, studentId, sStatus)
-      .then((data) => {
-        console.log("DATAAAAAAATRESSSSSSSS", data);
-        if (data.status >= 200) getStudentCohorts();
       })
       .catch((error) => error);
   };
@@ -149,200 +104,12 @@ const StudentCohorts = ({ stdId, setCohortOptions }) => {
     {
       disabled: false,
       component: (
-        <>
-          <Cohorts />
-          <div className="overflow-auto">
-          <div className="min-w-600">
-            {stdCohorts.map((s, i) => (
-              <div key={s.id} className="py-4">
-                <Grid container alignItems="center">
-                  <Grid item lg={4} md={4} sm={6} xs={6}>
-                    <div className="flex">
-                      <div className="flex-grow">
-                        <Link to={`/admissions/cohorts/${s.cohort.slug}`}>
-                          <h6 className="mt-0 mb-0 text-15 text-primary">
-                            {s.cohort.name}
-                            <small className="border-radius-4 ml-2 px-1 pt-2px bg-dark">
-                              {s.cohort.stage}
-                            </small>
-                          </h6>
-                        </Link>
-                        <p className="mt-0 mb-6px text-13">
-                          <span className="font-medium">
-                            {dayjs(s.cohort.kickoff_date).format(
-                              "DD MMMM, YYYY"
-                            )}
-                          </span>
-                          <small>
-                            , {dayjs(s.cohort.kickoff_date).fromNow()}
-                          </small>
-                        </p>
-                        <p className="mt-0 mb-6px text-13">
-                          <small
-                            onClick={() => {
-                              setRoleDialog(true);
-                              setCurrentStd({
-                                id: s.user.id,
-                                positionInArray: i,
-                                action: "role",
-                              });
-                            }}
-                            onKeyDown={() => {
-                              setRoleDialog(true);
-                              setCurrentStd({
-                                id: s.user.id,
-                                positionInArray: i,
-                                action: "role",
-                              });
-                            }}
-                            role="none"
-                            className="border-radius-4 px-2 pt-2px bg-secondary"
-                            style={{ cursor: "pointer", margin: "0 3px" }}
-                          >
-                            {s.role}
-                          </small>
-                          <small
-                            onClick={() => {
-                              setRoleDialog(true);
-                              setCurrentStd({
-                                id: s.user.id,
-                                positionInArray: i,
-                                action: "finantial_status",
-                              });
-                            }}
-                            onKeyDown={() => {
-                              setRoleDialog(true);
-                              setCurrentStd({
-                                id: s.user.id,
-                                positionInArray: i,
-                                action: "finantial_status",
-                              });
-                            }}
-                            role="none"
-                            className="border-radius-4 px-2 pt-2px bg-secondary"
-                            style={{ cursor: "pointer", margin: "0 3px" }}
-                          >
-                            {s.finantial_status ? s.finantial_status : "NONE"}
-                          </small>
-                          <small
-                            onClick={() => {
-                              setRoleDialog(true);
-                              setCurrentStd({
-                                id: s.user.id,
-                                positionInArray: i,
-                                action: "educational_status",
-                              });
-                            }}
-                            onKeyDown={() => {
-                              setRoleDialog(true);
-                              setCurrentStd({
-                                id: s.user.id,
-                                positionInArray: i,
-                                action: "educational_status",
-                              });
-                            }}
-                            role="none"
-                            className="border-radius-4 px-2 pt-2px bg-secondary"
-                            style={{ cursor: "pointer", margin: "0 3px" }}
-                          >
-                            {s.educational_status}
-                          </small>
-                          <small
-                            onClick={() => {
-                              setRoleDialog(true);
-                              setCurrentStd({
-                                id: s.user.id,
-                                positionInArray: i,
-                                action: "subscriptions_status",
-                              });
-                            }}
-                            onKeyDown={() => {
-                              setRoleDialog(true);
-                              setCurrentStd({
-                                id: s.user.id,
-                                positionInArray: i,
-                                action: "subscriptions_status",
-                              });
-                            }}
-                            role="none"
-                            className="border-radius-4 px-2 pt-2px bg-secondary"
-                            style={{
-                              cursor: "pointer",
-                              margin: "0 3px",
-                              color: "white",
-                            }}
-                          >
-                            {subscriptionSlugs[i]?.plan?.toUpperCase() ||
-                              "SUBSCRIPTION SLUG"}
-                          </small>
-
-                          <small
-                            onClick={() => {
-                              setRoleDialog(true);
-                              setCurrentStd({
-                                id: s.user.id,
-                                positionInArray: i,
-                                action: "plan_financing_status",
-                              });
-                            }}
-                            onKeyDown={() => {
-                              setRoleDialog(true);
-                              setCurrentStd({
-                                id: s.user.id,
-                                positionInArray: i,
-                                action: "plan_financing_status",
-                              });
-                            }}
-                            role="none"
-                            className="border-radius-4 px-2 pt-2px bg-secondary"
-                            style={{
-                              cursor: "pointer",
-                              margin: "0 3px",
-                              color: "red",
-                            }}
-                          >
-                            {planFinancingSlugs[i]?.plan?.toUpperCase() ||
-                              "PLAN FINANCING SLUG"}
-                          </small>
-                        </p>
-                      </div>
-                    </div>
-                  </Grid>
-                </Grid>
-              </div>
-            ))}
-          </div>
-        </div>
-
-        <Dialog
-          onClose={() => setRoleDialog(false)}
-          open={openRoleDialog}
-          aria-labelledby="simple-dialog-title"
-        >
-          <DialogTitle id="simple-dialog-title">{`Select a ${
-            actionController.message[currentStd.action]
-          }`}</DialogTitle>
-          <List>
-            {currentStd.action &&
-              actionController.options[currentStd.action].map((opt) => (
-                <ListItem
-                  button
-                  onClick={() => {
-                    changeStudentStatus(
-                      opt,
-                      currentStd.action,
-                      currentStd.id,
-                      currentStd.positionInArray
-                    );
-                    setRoleDialog(false);
-                  }}
-                >
-                  <ListItemText primary={opt} />
-                </ListItem>
-              ))}
-          </List>
-        </Dialog>
-        </>
+        <Cohorts 
+          stdCohorts={stdCohorts} 
+          subscriptionSlugs={subscriptionSlugs}
+          planFinancingSlugs={planFinancingSlugs}
+          getStudentCohorts={getStudentCohorts}
+        />
       ),
       label: "Cohorts",
     },
@@ -365,6 +132,8 @@ const StudentCohorts = ({ stdId, setCohortOptions }) => {
         if (data.length > 0) {
           const slugs = data.map((subscription) => ({
             plans: subscription?.plans,
+            cohorts: subscription?.selected_cohort_set?.cohorts,
+            id: subscription?.id
           }));
           console.log("SLUGS SUBSCRIPTION", slugs);
           setSubscriptionSlugs(slugs);
